app/requirements.txt
aniso8601==9.0.1
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
Jinja2==3.0.3
mongoengine==0.24.0
pika==1.2.0
pymongo==4.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
